+++ src/utils/dataUtils.js
@@ -183,7 +183,9 @@ export default {
         let response = (await restUtils.callConnector("GET", "/api/requests"))["_embedded"].resources;
         let resources = [];
         for (let idsResource of response) {
-            resources.push(clientDataModel.convertIdsResource(idsResource));
+            let resource = clientDataModel.convertIdsResource(idsResource);
+            resource.remoteId = idsResource.remoteId.substring(idsResource.remoteId.lastIndexOf("/") + 1, idsResource.remoteId.length);
+            resources.push(resource);
         }
         return resources;
     },
@@ -361,6 +363,44 @@ export default {
         return statusClass;
     },
 
+    async deleteAllRoutesOfApp(appId) {
+        let appEndpoints = await this.getAppEndpoints(appId);
+        let appEndpointIds = [];
+        for (let appEndpoint of appEndpoints) {
+            appEndpointIds.push(this.getIdOfConnectorResponse(appEndpoint));
+        }
+        let routes = await this.getRoutes();
+        for (let route of routes) {
+            let routeId = this.getIdOfConnectorResponse(route);
+            if (this.routeContainsEndpoint(route, appEndpointIds)) {
+                this.deleteRoute(routeId);
+            } else {
+                let subRoutes = await this.getRouteSteps(routeId);
+                for (let subRoute of subRoutes) {
+                    if (this.routeContainsEndpoint(subRoute, appEndpointIds)) {
+                        this.deleteRoute(routeId);
+                        break;
+                    }
+                }
+            }
+        }
+    },
+
+    routeContainsEndpoint(route, endpointsIds) {
+        let contains = false;
+        for (let endpointId of endpointsIds) {
+            if (route.start !== undefined && route.start != null && route.start.id == endpointId) {
+                contains = true;
+                break;
+            }
+            if (route.end !== undefined && route.end != null && route.end.id == endpointId) {
+                contains = true;
+                break;
+            }
+        }
+        return contains;
+    },
+
     async startApp(appId) {
         try {
             let params = {
@@ -373,14 +413,20 @@ export default {
     },
 
     async stopApp(appId) {
+        let inUse = false;
         try {
             let params = {
                 "type": "STOP"
             };
             await restUtils.callConnector("PUT", "/api/apps/" + appId + "/actions", params);
         } catch (error) {
-            errorUtils.showError(error, "Stop app");
+            if (error.details !== undefined && error.details.data !== undefined && error.details.data.message == "Selected App is in use by Camel.") {
+                inUse = true;
+            } else {
+                errorUtils.showError(error, "Stop app");
+            }
         }
+        return inUse;
     },
 
     async isAppRunning(appId) {
@@ -397,6 +443,13 @@ export default {
     },
 
     async deleteApp(appId) {
+        if (await this.isAppRunning(appId)) {
+            await this.stopApp(appId);
+        }
+        let params = {
+            "type": "DELETE"
+        };
+        await restUtils.callConnector("PUT", "/api/apps/" + appId + "/actions", params);
         await restUtils.callConnector("DELETE", "/api/apps/" + appId);
     },
 
@@ -405,7 +458,7 @@ export default {
             "recipient": appStoreUrl,
             "appId": appUrl
         }
-        await restUtils.callConnector("POST", "/api/ids/app", params);
+        return await restUtils.callConnector("POST", "/api/ids/app", params);
     },
 
     async getAppsOfAppStore(appStoreUrl) {
@@ -542,6 +595,18 @@ export default {
         return await restUtils.callConnector("GET", "/api/datasources/" + id);
     },
 
+    async getSubscriptions() {
+        let subscriptions = (await restUtils.callConnector("GET", "/api/subscriptions"))._embedded.subscriptions;
+        for (let subscription of subscriptions) {
+            subscription.creationDate = subscription.creationDate.substring(0, 19).replace("T", " ");
+        }
+        return subscriptions;
+    },
+
+    async deleteSubscription(id) {
+        await restUtils.callConnector("DELETE", "/api/subscriptions/" + id);
+    },
+
     async getGenericEndpoints() {
         let genericEndpoints = [];
         let idsEndpoints = (await restUtils.callConnector("GET", "/api/endpoints"))._embedded.endpoints;
@@ -631,7 +696,7 @@ export default {
             "description": desc
         });
 
-        if (sourceType != "OTHER") {
+        if (sourceType !== "OTHER") {
             let bodyData = null;
             if (username != null) {
                 bodyData = {
@@ -652,7 +717,7 @@ export default {
                     "url": url
                 };
             }
-            if (sourceType == "DATABASE") {
+            if (sourceType === "DATABASE") {
                 bodyData.url = url;
                 bodyData.driverClassName = driverClassName;
             }
@@ -680,6 +745,10 @@ export default {
         await restUtils.callConnector("DELETE", "/api/artifacts/" + resource.artifactId);
     },
 
+    async deleteRequestedResource(id) {
+        await restUtils.callConnector("DELETE", "/api/requests/" + id);
+    },
+
     async getRoute(id) {
         return await restUtils.callConnector("GET", "/api/routes/" + id);
     },
@@ -689,6 +758,11 @@ export default {
     },
 
     async deleteRoute(id) {
+        let artifact = await restUtils.callConnector("GET", "/api/routes/" + id + "/output");
+        if (typeof artifact == 'object') {
+            let artifactId = this.getIdOfConnectorResponse(artifact);
+            await restUtils.callConnector("DELETE", "/api/artifacts/" + artifactId);
+        }
         await restUtils.callConnector("DELETE", "/api/routes/" + id);
     },
 
@@ -941,7 +1015,7 @@ export default {
                 await restUtils.callConnector("DELETE", "/api/catalogs/" + catalogId + "/offers", null, [resourceId]);
             }
 
-            // Delete all rules and create new ones. Rules that have not been edited in the UI are also deleted and recreated. 
+            // Delete all rules and create new ones. Rules that have not been edited in the UI are also deleted and recreated.
             // Implementing the detection of a change to an existing rule would have been complicated (Pattern can change, only value can change, ...)
 
             let contractId = -1;
@@ -1082,6 +1156,10 @@ export default {
         await restUtils.callConnector("PUT", "/api/routes/" + routeId + "/endpoint/start", null, "\"" + startSelfLink + "\"");
     },
 
+    async addRouteEnd(routeId, endSelfLink) {
+        await restUtils.callConnector("PUT", "/api/routes/" + routeId + "/endpoint/end", null, "\"" + endSelfLink + "\"");
+    },
+
     async addRouteStartAndEnd(routeId, startId, endId) {
         await restUtils.callConnector("PUT", "/api/routes/" + routeId + "/endpoint/start", null, "\"" + startId + "\"");
         await restUtils.callConnector("PUT", "/api/routes/" + routeId + "/endpoint/end", null, "\"" + endId + "\"");
@@ -1095,11 +1173,10 @@ export default {
         });
     },
 
-    async createSubRoute(routeId, startId, startCoordinateX, startCoordinateY, endId, endCoordinateX, endCoordinateY, artifactId) {
+    async createSubRoute(routeId, startSelfLink, startCoordinateX, startCoordinateY, endSelfLink, endCoordinateX, endCoordinateY) {
         let hasError = false;
         try {
             let response = await restUtils.callConnector("POST", "/api/routes", null, {
-                "routeType": "Subroute",
                 "deploy": "None",
                 "startCoordinateX": startCoordinateX,
                 "startCoordinateY": startCoordinateY,
@@ -1107,11 +1184,14 @@ export default {
                 "endCoordinateY": endCoordinateY
             });
             let subRouteId = this.getIdOfConnectorResponse(response);
-            await restUtils.callConnector("POST", "/api/routes/" + routeId + "/steps", null, [subRouteId]);
-            await restUtils.callConnector("PUT", "/api/routes/" + subRouteId + "/endpoint/start", null, "\"" + startId + "\"");
-            await restUtils.callConnector("PUT", "/api/routes/" + subRouteId + "/endpoint/end", null, "\"" + endId + "\"");
-            await restUtils.callConnector("POST", "/api/routes/" + subRouteId + "/outputs", null, [artifactId]);
-
+            let subRouteSelfLink = response._links.self.href;
+            if (startSelfLink !== undefined && startSelfLink != null) {
+                await restUtils.callConnector("PUT", "/api/routes/" + subRouteId + "/endpoint/start", null, "\"" + startSelfLink + "\"");
+            }
+            if (endSelfLink !== undefined && endSelfLink != null) {
+                await restUtils.callConnector("PUT", "/api/routes/" + subRouteId + "/endpoint/end", null, "\"" + endSelfLink + "\"");
+            }
+            await restUtils.callConnector("POST", "/api/routes/" + routeId + "/steps", null, [subRouteSelfLink]);
         } catch (error) {
             errorUtils.showError(error, "Save Route");
             hasError = true;
@@ -1258,7 +1338,7 @@ export default {
             let response = await restUtils.callConnector("POST", "/api/ids/description", params);
             return response;
         },
-    
+
         async receiveIdsContractOffer(recipientId, artifactId) {
             let params = {
                 "recipient": recipientId,
@@ -1282,7 +1362,7 @@ export default {
         return await restUtils.callConnector("POST", "/api/ids/contract", params, contractOffer[0]["ids:permission"]);
     },
 
-    async subscribeToResource(recipientId, resoureceId) {
+    async subscribeToResource(recipientId, resoureceId, subscriptionLocation) {
         let params = {
             "recipient": recipientId,
         }
@@ -1293,7 +1373,7 @@ export default {
             "title": "default",
             "description": "Notify on update",
             "target": resoureceId,
-            "location": configuration.endpoint,
+            "location": subscriptionLocation,
             "subscriber": configuration.id,
             "pushData": true
         }